Refactor out list-processes command handler and add it to dotnet-dump (#534)
* Refactor out ps command handler and add it to dotnet-dump

* More refactoring

* Some renaming

* update doc

* More docs update

* fix broken renamed file reference

* catch InvalidOperationException only
