+++ documentation/dotnet-counters-instructions.md
@@ -32,7 +32,19 @@ dotnet tool install --global dotnet-counters
 *COMMANDS*
 
     list      Display a list of counter names and descriptions
+    ps        Display a list of dotnet processes that can be monitored
     monitor   Display periodically refreshing values of selected counters
+
+*PS*
+    dotnet-counters ps
+
+    Display a list of dotnet processes that can be monitored.
+
+    Examples:
+      > dotnet-counters ps
+
+     15683 WebApi     /home/suwhang/repos/WebApi/WebApi
+     16324 dotnet     /usr/local/share/dotnet/dotnet
 
 *LIST*
